pyproject.toml
[project]
name = "fdds-rag"
version = "0.1.0"
description = "A Python package for generating AI-powered responses using retrieval-augmented generation."
requires-python = ">=3.11"
dependencies = [
    "pre-commit>=4.1.0",
    "pydantic-settings>=2.8.1",
    "ragbits[qdrant]==1.3.0",
    "ragbits-evaluate[relari]>=1.3.0",
    "unstructured[pdf]>=0.17.2",
    "fastapi[standard]>=0.115.11",
    "scrapy>=2.12.0",
    "opentelemetry-sdk>=1.32.0",
    "opentelemetry-exporter-otlp>=1.32.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fdds"]
